Moved the core_system default values file into client module and removed all references to it in api.py
